[REF] *_loyalty_partler_applicability: No restriction on partner at rule level
At the request of @pilarvargas-tecnativa, removal of the option to limit the applicability of a rule to certain partners only. To simplify the code, the restriction can now only be defined at the programme level.
