+++ loyalty_partner_applicability/models/__init__.py
@@ -1,4 +1,2 @@
-from . import loyalty_partner_applicability_mixin
 from . import loyalty_program
-from . import loyalty_rule
 from . import res_config_settings
