+++ loyalty_partner_applicability/static/description/index.html
@@ -374,15 +374,9 @@ ul.auto-toc {
 filter for promotion rules. This module is a base to be extended.</p>
 <p>It allows to restrict the applicability of a promotion and its rules to a subset of customers.
 This is useful when you want to target a specific group of customers with a promotion.
-The restriction can be defined at 2 levels:</p>
-<ul class="simple">
-<li>The program</li>
-<li>The rule</li>
-</ul>
+The restriction can be defined at the program level.</p>
 <p>When a restriction is defined at the program level, the promotion will only be available to
-customers that match the filter. When a restriction is defined at the rule level, the rule
-will only be applied to customers that match the filter. If a restriction is defined at both
-levels, the program’s rules will only be applied to customers that match both filters.</p>
+customers that match the filter.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
