--- loyalty_partner_applicability/views/loyalty_rule_view_form.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="loyalty_rule_view_form" model="ir.ui.view">
-        <field name="inherit_id" ref="loyalty.loyalty_rule_view_form" />
-        <field name="model">loyalty.rule</field>
-        <field name="arch" type="xml">
-            <field name="minimum_qty" position="before">
-                <field
-                    name="partner_ids"
-                    widget="many2many_tags"
-                    options="{'no_create': True}"
-                />
-                <field
-                    name="partner_domain"
-                    widget="domain"
-                    options="{'model': 'res.partner', 'in_dialog': true}"
-                />
-            </field>
-        </field>
-    </record>
-</odoo>
